Use more reliable path detection logic
In case CUDA_HOME is not specified, lib64 was assumed to be the location of
libcudart.so. This is a rather crude assumption and fails on 32 bit and
multi-arch distributions like Debian/Ubuntu. This change uses the ldconfig cache
to determine the correct path and works successfully on a Ubuntu 16.04 machine.